requirements.txt
bottle==0.13.1
bottle-websocket==0.2.9
Eel==0.17.0
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.1.1
mysql-connector-python==9.0.0
pyparsing==3.1.4
whichcraft==0.6.1
zope.event==5.0
zope.interface==7.0.3
